Closes #30 Add check for file existence

+++ dbeaver/install.sh
@@ -6,15 +6,18 @@ source $(dirname $0)/../secrets/dots.sh
 set -e
 
 function move_dbeaver_connections_from() {
-  mv ~/.dbeaver4/General/.dbeaver-data-sources.xml  ~/.dbeaver4/General/.dbeaver-data-sources.xml.original
+  if [ -f ~/.dbeaver4/General/.dbeaver-data-sources.xml ]
+  then
+    mv ~/.dbeaver4/General/.dbeaver-data-sources.xml  ~/.dbeaver4/General/.dbeaver-data-sources.xml.original
+  fi
   cp $1/dbeaver/dbeaver-data-sources.xml ~/.dbeaver4/General/.dbeaver-data-sources.xml
 }
+
+setup_secret_dotfiles;
+
+move_dbeaver_connections_from "$HOME/.dot-secrets";
 
 if test ! $(pgrep -f "DBeaver")
 then
   open "/Applications/DBeaver.app"
 fi
-
-setup_secret_dotfiles;
-
-move_dbeaver_connections_from "$HOME/.dot-secrets";
